Added warning if signals are clipping on low side
Added warning if signals are clipping on the low side, i.e. if they are saturated at zero volts.  As we expect signals very near 0 volts when the LEDs are off the clipping warning is only shown if the signal is at exactly 0 ADC reading for multiple samples, which should only happen if the input is pulled below 0V.